+++ iosApp/iosApp/viewModel/reports/model/ReportModel.swift
@@ -3,7 +3,7 @@ import Foundation
 struct ReportModel: Identifiable {
     let id: Int64
     let title: String
-    let color: Int32
+    let color: Int64
     let proportion: Double
     
 }
